feat: add api userWord
