+++ src/api/api.ts
@@ -1,13 +1,12 @@
-import { IWord, IUser, IUpdateUser, IUserWord, ISignIn, SignIn, ICreateUser } from '../interfaces/interfaces'
-
-const baseUrl = 'https://team83-rslang.herokuapp.com';
+import { IWord, IUser, IUpdateUser, IUserWord, ISignIn, SignIn, ICreateUser, IFetchOptions } from '../interfaces/interfaces'
+import { BASEURL, RESPONSE } from './consts';
 
 export const getWords = async (group = 0, page = 0,): Promise<{ words: Array<IWord> } | null> => {
   try {
-    const data = await fetch(`${baseUrl}/words?group=${group}&page${page}`);
+    const data = await fetch(`${BASEURL}/words?group=${group}&page${page}`);
     const res: IWord[] = await data.json()
 
-    if (data.status === 200) { // TODO in this folder create consts.ts and describe 200 and other codes
+    if (data.status === RESPONSE.OK) { // TODO in this folder create consts.ts and describe 200 and other codes
       return {
         words: res
       }
@@ -20,10 +19,10 @@ export const getWords = async (group = 0, page = 0,): Promise<{ words: Array<IWo
 
 export const getWord = async (wordID: string): Promise<IWord | null> => {
   try {
-    const data = await fetch(`${baseUrl}/words/${wordID}`);
+    const data = await fetch(`${BASEURL}/words/${wordID}`);
     const res: IWord = await data.json()
 
-    if (data.status === 200) {
+    if (data.status === RESPONSE.OK) {
       return res
     }
     return null
@@ -32,68 +31,8 @@ export const getWord = async (wordID: string): Promise<IWord | null> => {
   }
 }
 
-export const getUser = async (UserID: string): Promise<IUser | null> => {
-  try {
-    const data = await fetch(`${baseUrl}/users/${UserID}`);
-    const res: IUser = await data.json()
-
-    if (data.status === 200) {
-      return res
-    }
-    if (data.status === 404) {
-      return res // TO DO  User not found
-    }
-    if (data.status === 401) {
-      return res // TO DO  	Access token is missing or invalid
-    }
-    return null;
-
-  } catch (err) {
-    throw new Error("Error");
-  }
-}
-export const createUser = async (user: ICreateUser): Promise<void> => {
-  try {
-    await fetch(`${baseUrl}/users`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(user),
-    });
-
-  }
-  catch (err) {
-    throw new Error("Error");
-  }
-}
-
-export const deleteUser = async (userID: string): Promise<void> => {
-  try {
-    await fetch(`${baseUrl}/user/${userID}`, {
-      method: 'DELETE',
-    });
-  }
-  catch (err) {
-    throw new Error("Error");
-  }
-}
-export const updateUser = async (user: IUpdateUser): Promise<void> => {
-  try {
-    await fetch(`${baseUrl}/user/${user.id}`, {
-      method: 'PUT',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(user),
-    });
-  } catch (err) {
-    throw new Error(err);
-  }
-};
-
 export const signIn = async (user: SignIn): Promise<ISignIn | null> => {
-  const data = await fetch(`${baseUrl}/signin`,
+  const data = await fetch(`${BASEURL}/signin`,
     {
       method: 'POST',
       headers: {
@@ -103,15 +42,55 @@ export const signIn = async (user: SignIn): Promise<ISignIn | null> => {
       body: JSON.stringify(user)
     });
   const res = await data.json();
-  if (data.status === 200) {
-    saveToken(res.token);
+  if (data.status === RESPONSE.OK) {
+    saveToken(res.token, res.refreshToken, res.userId);
     return res
   }
-  return null;
+  if (data.status === RESPONSE.Uncorrect) {
+    return null
+  }
+  return null
 }
-function saveToken(token: string) {
-  sessionStorage.setItem('tokenData', token);
+function saveToken(token: string, refreshToken: string, userId: string) {
+  sessionStorage.setItem('token', token);
+  sessionStorage.setItem('refreshToken', refreshToken);
+  sessionStorage.setItem('tokenDate', Date.now().toString())
+  sessionStorage.setItem('userId', userId);
   sessionStorage.setItem('isAutorization', 'true');
 }
 
+export const getRefreshToken = async (userId: string, refreshToken: string) => {
+  const data = await fetch(`${BASEURL}/users/${userId}/tokens`, {
+    method: 'GET',
+    /*withCredentials: true,*/
+    headers: {
+      'Authorization': `Bearer ${refreshToken}`,
+      'Accept': 'application/json',
+      'Content-Type': 'application/json'
+    }
+  });
+  const content = await data.json();
+  saveToken(content.token, content.refreshToken, userId)
+  return content.token
+}
+
+export const fetchWithAutorization = async (url: string, options?: IFetchOptions) => {
+  const token = sessionStorage.getItem('token');
+  const refreshToken = sessionStorage.getItem('refreshToken');
+  const userId = sessionStorage.getItem('userId');
+  const tokenDate = sessionStorage.getItem('tokenDate');
+  console.log(Date.now() - (Number(tokenDate)))
+  if (Date.now() - (Number(tokenDate)) <= 3600) {
+    if (userId && refreshToken) {
+      const newToken = getRefreshToken(userId, refreshToken)
+      if (options) {
+        options.headers.Authorization = `Bearer ${newToken}`;
+        console.log('NewToken', token)
+      } else {
+        const newToken = token;
+      }
+    }
+  }
+  return fetch(url, options);
+}
 //TODO methods post, put,etc. need validate
