+++ src/api/consts.ts
@@ -1,0 +1,5 @@
+export const BASEURL = 'https://team83-rslang.herokuapp.com';
+export const RESPONSE = {
+  'OK': 200,
+  'Uncorrect': 309,
+}
