+++ src/components/gameSprint/gameSprint.css
